refactor(filesystem): remove useless file
